requirements.txt
# Core NLP
spacy>=3.7.0
fastcoref>=2.1.0

# PDF Processing
PyPDF2>=3.0.0

# Graph & Visualization
networkx>=3.0
pyvis>=0.3.0

# ML/Embeddings (optional, for clustering features)
# sentence-transformers>=2.0.0
# scikit-learn>=1.0.0

# OpenAI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.0.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# spaCy model (install separately)
# python -m spacy download en_core_web_sm
